dto 각 클래스별로 분리
