+++ src/customers/dto/update-customer.dto.ts
@@ -1,0 +1,36 @@
+import { IsString, IsEnum, IsEmail, IsOptional, Length, IsInt, Min } from 'class-validator';
+import { Transform } from 'class-transformer';
+import { Gender, AgeGroup, Region } from '@prisma/client';
+
+export class UpdateCustomerDto {
+  @IsOptional()
+  @IsString()
+  @Length(1, 50)
+  name?: string;
+
+  @IsOptional()
+  @IsEnum(Gender)
+  gender?: Gender;
+
+  @IsOptional()
+  @IsString()
+  @Length(1, 20)
+  phoneNumber?: string;
+
+  @IsOptional()
+  @IsEnum(AgeGroup)
+  ageGroup?: AgeGroup;
+
+  @IsOptional()
+  @IsEnum(Region)
+  region?: Region;
+
+  @IsOptional()
+  @IsEmail()
+  @Length(1, 100)
+  email?: string;
+
+  @IsOptional()
+  @IsString()
+  memo?: string;
+}
